+++ template/cn/shoukaiseki/tmpl/SKS_TMPL_DATABEAN.js
@@ -147,6 +147,38 @@ function _close(set) {
   } catch (ignored) { }
 }
 
+
+/**
+ * 获取mbo
+ * @param {psdi.webclient.system.beans.DataBeanContext} dbctx - 数据Bean上下文
+ * @return {psdi.mbo.MboRemote}   主Mbo
+ */
+function getMainMbo(dbctx) {
+  logger.info("[" + scriptName + "] getMainMbo")
+
+  /** @type {psdi.webclient.system.controller.AppInstance} */
+  var appInstance = dbctx.getAppInstance();
+  /** @type {psdi.webclient.system.session.WebClientSession} */
+  var clientsession = dbctx.webclientsession();
+  /** @type {psdi.webclient.system.beans.DataBean} */
+  var appBean = appInstance.getAppBean();
+  /** @type {psdi.mbo.MboRemote} */
+  var mbo = appBean.getMbo();
+  if (!mbo) {
+    var appInstance = dbctx.getAppInstance()
+    logger.info("[" + scriptName + "] appInstance= " + appInstance)
+    var appBean = appInstance.getAppBean()
+    logger.info("[" + scriptName + "] appBean= " + appBean)
+    //应用主列表按钮的mbo获取
+    mbo = appBean.getMbo()
+    logger.info("[" + scriptName + "] mbo= " + mbo)
+  }
+  if (!mbo) {
+    throw new MXApplicationException("#", "error,can't find mbo")
+  }
+  return mbo;
+}
+
 /**
  * 一定要DATABEAN. 开头
  *